requirements.txt
appdirs==1.4.3
cycler==0.10.0
decorator==4.0.11
functools32==3.2.3.post2
matplotlib==2.0.2
networkx==1.11
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-mnist==0.3
pytz==2017.2
PyWavelets==0.5.2
scikit-image==0.13.0
scipy==0.19.0
six==1.10.0
subprocess32==3.2.7
